projects/practice-project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "de-project"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["pyyaml"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"de_project.config" = ["*.yaml"]
